WIP working on handling page close
