+++ pages/_app.js
@@ -1,4 +1,4 @@
-import { useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import PropTypes from 'prop-types'
 // NOTE: If I change the import below to go to @mui
 // the colors on the header are lost!?
@@ -18,6 +18,7 @@ import theme from '../src/theme'
 import '@styles/globals.css'
 import NonSSRWrapper from '@components/NonSSRWrapper'
 
+
 export default function Application({ Component, pageProps }) {
   useEffect(() => {
     // Remove the server-side injected CSS.
@@ -27,6 +28,44 @@ export default function Application({ Component, pageProps }) {
       jssStyles.parentElement.removeChild(jssStyles)
     }
   }, [])
+
+  const [firstTime, setFirstTime] = useState(true)
+  useEffect( () => {
+    if ( firstTime ) {
+      // hopefully this only runs one time...
+      sessionStorage.setItem("unsavedChanges", 0)
+      setFirstTime(false)
+    }
+  }, [firstTime])
+
+  // prompt the user if they try and leave with unsaved changes  
+  useEffect(() => {
+    const warningText =
+      'You have unsaved changes - are you sure you wish to leave this page?';
+    const handleWindowClose = (e) => {    
+      let unsavedChanges = false
+      let _unsavedChanges = sessionStorage.getItem("unsavedChanges")
+      console.log("_unsavedChanges type,value",typeof _unsavedChanges, _unsavedChanges)
+      if ( _unsavedChanges === "" || _unsavedChanges === null || _unsavedChanges === "0" ) { 
+        return
+      }
+      // if (!_unsavedChanges) return;
+      e.preventDefault();
+      return (e.returnValue = warningText);
+    };
+    // const handleBrowseAway = () => {
+    //   if (!unsavedChanges) return;
+    //   if (window.confirm(warningText)) return;
+    //   router.events.emit('routeChangeError');
+    //   throw 'routeChange aborted.';
+    // };
+    window.addEventListener('beforeunload', handleWindowClose);
+    // router.events.on('routeChangeStart', handleBrowseAway);
+    return () => {
+      window.removeEventListener('beforeunload', handleWindowClose);
+      // router.events.off('routeChangeStart', handleBrowseAway);
+    };
+  }, []);
 
   return (
     <div>
